# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Cupreous Dating App"
    },
    "host": "http://54.152.239.30/:9118",
    "basePath": "/",
    "paths": {
...
        "/user/forgot": {
            "post": {
                "tags": ["user"],
                "summary": "This is to Recover Forgotten Password",
                "description": "API for getting forgotten Password.",
                "operationId": "forgotPassword",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "email",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ForgotPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                }
            }
        },
...
        "/user/question/{id}": {
            "put": {
                "tags": ["questionarieApi"],
                "summary": "Update questions",
                "description": "This can only be done by the logged in user.",
                "operationId": "updateQuestion",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "questions that need to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateQuestion"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Questions not found"
                    }
                }
            },
            "delete":{
                "tags": ["questionarieApi"],
                "summary": "Delete questions",
                "description": "This can only be done by the logged in user.",
                "operationId": "delete",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "question that need to be delete",
                        "required": true,
                        "type": "string"
                    }                    
                ],
                "responses": {
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Questions not found"
                    }
                }
            }
        },
...
    },
...
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "contactNo": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                
                "gender": {
                    "type": "string"
                },
                "height": {
                    "type": "string"
                },
               
                "maritalStatus": {
                    "type": "string"
                },
                "qualification": {
                    "type": "string"
                },
                
                "password": {
                    "type": "String"
                },
                "confirmPassword": {
                    "type": "string"
                },
                
                "ethnicity": {
                    "type": "string"
                }
               
            }
        },
...
        "ForgotPassword": {
            "type": "object",
            "properties": {
                "email": {
                    "type":"string"
                }
            }
        },
...
        "updateQuestion": {
            "type": "object",
            "properties": {
                "questionNo": {
                    "type": "string"
                }
            }
        },
...
    },
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"]
}